requirements.txt
bidict
cidict
pathlib
pymills
procname
circuits
python-magic
